# openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /api/contacts/:
    post:
      tags:
        - Contact manager
      description: Saves the client's contact in DB, if this contact is not represented in the DB
      operationId: createContact
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/contact'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json: {}
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
  /api/clients:
    get:
      tags:
        - Client manager
      description: Returns list of all clients name in DB
      operationId: getClients
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/client'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
    post:
      tags:
        - Client manager
      description: Saves the client's name in DB, if this client's name is not represented in DB
      operationId: createClient
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/client'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json: {}
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
  /openapi.yaml:
    get:
      tags:
        - OpenApi manager
      description: Returns OpenApi specification
      operationId: getOpenApiYaml
      responses:
        '200':
          description: OK
          content:
            application/json: {}
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
  /api/contacts/{id}:
    get:
      tags:
        - Contact manager
      description: Returns list of contact by id. If the optional parameter 'type' is specified in the query string, returns list of contacts of the specified type(email/phone)
      operationId: getContacts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: type
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contact'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
  /api/clients/{id}:
    get:
      tags:
        - Client manager
      description: Returns client name by id
      operationId: getClientNameById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/client'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
components:
  schemas:
    contact:
      required:
        - type
        - value
      type: object
      properties:
        type:
          type: string
          enum:
            - EMAIL
            - PHONE
        value:
          type: string
      description: Simple contact info.
    client:
      required:
        - name
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
      description: Client name must be unique